move Pytorch model builder together with base pytorch model
