+++ pyzoo/zoo/automl/model/base_pytorch_model.py
@@ -18,6 +18,7 @@ from torch.utils.data import TensorDataset, DataLoader
 
 import types
 
+from zoo.automl.model import ModelBuilder
 from zoo.automl.model.abstract import BaseModel
 from zoo.automl.common.util import *
 from zoo.automl.common.metrics import Evaluator
@@ -347,3 +348,31 @@ class PytorchBaseModel(BaseModel):
                 "optim",
                 "loss"
                 }
+
+
+class PytorchModelBuilder(ModelBuilder):
+
+    def __init__(self, model_creator,
+                 optimizer_creator,
+                 loss_creator):
+        from zoo.orca.automl.pytorch_utils import validate_pytorch_loss, validate_pytorch_optim
+        self.model_creator = model_creator
+        optimizer = validate_pytorch_optim(optimizer_creator)
+        self.optimizer_creator = optimizer
+        loss = validate_pytorch_loss(loss_creator)
+        self.loss_creator = loss
+
+    def build(self, config):
+        model = PytorchBaseModel(self.model_creator,
+                                 self.optimizer_creator,
+                                 self.loss_creator)
+        model.build(config)
+        return model
+
+    def build_from_ckpt(self, checkpoint_filename):
+        '''Restore from a saved model'''
+        model = PytorchBaseModel(self.model_creator,
+                                 self.optimizer_creator,
+                                 self.loss_creator)
+        model.restore(checkpoint_filename)
+        return model
