+++ pyzoo/zoo/chronos/autots/model/auto_lstm.py
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from zoo.automl.model import PytorchModelBuilder
+from zoo.automl.model.base_pytorch_model import PytorchModelBuilder
 from zoo.orca.automl.auto_estimator import AutoEstimator
 from zoo.chronos.model.VanillaLSTM_pytorch import model_creator
